feat(conn): native cryptoutils
